requirements.txt
Django==3.0.5
django-crispy-forms==1.9.0
django-import-export==2.0.2
django-recaptcha==2.0.6
django-wkhtmltopdf==3.3.0
mysqlclient==1.4.6
social-auth-app-django==3.1.0
uWSGI==2.0.18